requirements.txt
# Core dependencies for Resume Parser with Batch Processing
google-generativeai>=0.3.0
redis>=5.0.0
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0

# File processing libraries
python-docx>=0.8.11
docx2txt>=0.8
PyPDF2>=3.0.0